requirements.txt
anyio==3.5.0
fastapi==0.75.0
idna==3.3
pydantic[email]==1.9.0
sniffio==1.2.0
SQLAlchemy==1.4.32
sqlalchemy2-stubs==0.0.2a20
sqlmodel==0.0.6
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
python-dotenv==0.19.2
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4